+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/AIButton.tsx
@@ -1,3 +1,4 @@
+import { Block } from '@blocknote/core';
 import {
   ComponentProps,
   useBlockNoteEditor,
@@ -225,14 +226,23 @@ const AIMenuItemTransform = ({
   icon,
 }: PropsWithChildren<AIMenuItemTransform>) => {
   const { mutateAsync: requestAI, isPending } = useDocAITransform();
+  const editor = useBlockNoteEditor();
 
-  const requestAIAction = async (markdown: string) => {
+  const requestAIAction = async (selectedBlocks: Block[]) => {
+    const text = await editor.blocksToMarkdownLossy(selectedBlocks);
+
     const responseAI = await requestAI({
-      text: markdown,
+      text,
       action,
       docId,
     });
-    return responseAI.answer;
+
+    if (!responseAI || !responseAI.answer) {
+      throw new Error('No response from AI');
+    }
+
+    const markdown = await editor.tryParseMarkdownToBlocks(responseAI.answer);
+    editor.replaceBlocks(selectedBlocks, markdown);
   };
 
   return (
@@ -255,14 +265,39 @@ const AIMenuItemTranslate = ({
   language,
 }: PropsWithChildren<AIMenuItemTranslate>) => {
   const { mutateAsync: requestAI, isPending } = useDocAITranslate();
+  const editor = useBlockNoteEditor();
 
-  const requestAITranslate = async (markdown: string) => {
+  const requestAITranslate = async (selectedBlocks: Block[]) => {
+    console.log('selectedBlocks', selectedBlocks);
+
+    let fullHtml = '';
+    for (const block of selectedBlocks) {
+      console.log('block', block.content);
+      if (Array.isArray(block.content) && block.content.length === 0) {
+        fullHtml += '<p><br/></p>';
+        continue;
+      }
+
+      fullHtml += await editor.blocksToHTMLLossy([block]);
+    }
+    console.log('html', fullHtml);
+
     const responseAI = await requestAI({
-      text: markdown,
+      text: fullHtml,
       language,
       docId,
     });
-    return responseAI.answer;
+
+    if (!responseAI || !responseAI.answer) {
+      throw new Error('No response from AI');
+    }
+
+    try {
+      const blocks = await editor.tryParseHTMLToBlocks(responseAI.answer);
+      editor.replaceBlocks(selectedBlocks, blocks);
+    } catch {
+      editor.replaceBlocks(selectedBlocks, selectedBlocks);
+    }
   };
 
   return (
@@ -277,7 +312,7 @@ const AIMenuItemTranslate = ({
 };
 
 interface AIMenuItemProps {
-  requestAI: (markdown: string) => Promise<string>;
+  requestAI: (blocks: Block[]) => Promise<void>;
   isPending: boolean;
   icon?: ReactNode;
 }
@@ -293,31 +328,17 @@ const AIMenuItem = ({
   const editor = useBlockNoteEditor();
   const handleAIError = useHandleAIError();
 
-  const handleAIAction = async () => {
+  const handleAIAction = () => {
     let selectedBlocks = editor.getSelection()?.blocks;
 
     if (!selectedBlocks || selectedBlocks.length === 0) {
       selectedBlocks = [editor.getTextCursorPosition().block];
-
       if (!selectedBlocks || selectedBlocks.length === 0) {
         return;
       }
     }
 
-    const markdown = await editor.blocksToMarkdownLossy(selectedBlocks);
-
-    try {
-      const responseAI = await requestAI(markdown);
-
-      if (!responseAI) {
-        return;
-      }
-
-      const blockMarkdown = await editor.tryParseMarkdownToBlocks(responseAI);
-      editor.replaceBlocks(selectedBlocks, blockMarkdown);
-    } catch (error) {
-      handleAIError(error);
-    }
+    requestAI(selectedBlocks).catch(handleAIError);
   };
 
   if (!Components) {
@@ -351,6 +372,7 @@ const useHandleAIError = () => {
       return;
     }
 
+    console.error('AI', error);
     toast(t('AI seems busy! Please try again.'), VariantType.ERROR);
   };
 };